Add a mel_sorter to the library.
First melsort draft. It's not optimized and I fear it also doesn't follow
the original description of the algorithm (no binary search, the tricks to
save memory seem to hint at a different implementation too). Also it's
currently very slow, but maybe it's impossible to make it a truly
efficient sort.